fix: 컴포넌트 재분리 및 emotion 설치
