+++ package.json
@@ -3,6 +3,8 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@emotion/react": "^11.11.1",
+    "@emotion/styled": "^11.11.0",
     "@testing-library/user-event": "^13.5.0",
     "@types/node": "^16.18.46",
     "react": "^18.2.0",
